defect/10-container-image-build-fails-with-missing-diffpy-error
Closes #10

* Jupyter support removed

docker/Dockerfile
* ENTRYPOINT refactor so that diffit container can support all
  subcommands

Makefile
* docker buildx support
* new target "diffit-image-build" to handle package ZIP and container
  image build
