+++ Makefile
@@ -7,7 +7,9 @@ include makester/makefiles/makester.mk
 MAKESTER__PACKAGE_NAME := diffit
 
 ifeq ($(MAKESTER__ARCH), arm64)
-  DOCKER_PLATFORM := --platform linux/amd64
+DOCKER_PLATFORM := linux/arm64/v8
+else
+DOCKER_PLATFORM := linux/amd64
 endif
 
 MAKESTER__WHEEL := .wheelhouse
@@ -37,53 +39,47 @@ tests:
  -p tests.diffit.dataframes\
  --junitxml junit.xml $(TESTS)
 
-# Update the README if changing the default Spark version.
-SPARK_VERSION := 3.2.1
-# Tagging convention used: <spark-version>-<workflow-version>-<image-release-number>
-MAKESTER__VERSION := $(SPARK_VERSION)-$(RELEASE_VERSION)
+# Tagging convention used: <spark-version>-<diffit-version>-<image-release-number>
 MAKESTER__RELEASE_NUMBER ?= 1
 
 # Tag the image build
-export MAKESTER__IMAGE_TAG_ALIAS = $(MAKESTER__SERVICE_NAME):$(MAKESTER__RELEASE_VERSION)-$(MAKESTER__RELEASE_NUMBER)
+export MAKESTER__IMAGE_TAG_ALIAS := $(MAKESTER__SERVICE_NAME):$(MAKESTER__RELEASE_VERSION)-$(MAKESTER__RELEASE_NUMBER)
 
-py-distribution: MAKESTER__WHEEL = .wheelhouse
+py-distribution: MAKESTER__WHEEL := .wheelhouse
 
 lint:
 	-@pylint $(MAKESTER__PROJECT_DIR)/src
 
-dep-builder: APP_ENV = prod
-dep-builder: PIP_INSTALL = --upgrade --target .dependencies .
-dep-builder: py-venv-init
+dep-builder:
+	$(MAKESTER__PIP) install --upgrade --target .dependencies .
 
-dep-package: dep-builder
+dep-package: _dep-package-setup dep-builder _dep-package
+_dep-package-setup:
 	$(info ### Building the spark-submit python dependencies zip file)
-	@$(shell which mkdir) -pv $(MAKESTER__PROJECT_DIR)/.dependencies\
- && cd $(MAKESTER__PROJECT_DIR)/.dependencies\
- && zip -r $(MAKESTER__PROJECT_DIR)/docker/files/python/dependencies.zip . -i *
+	@$(shell which mkdir) -pv $(MAKESTER__PROJECT_DIR)/.dependencies
+_dep-package:
+	cd $(MAKESTER__PROJECT_DIR)/.dependencies && zip -r $(MAKESTER__PROJECT_DIR)/docker/files/python/dependencies.zip .
 
 CMD ?= --help
 diffit:
 	@diffit $(CMD)
 
-differ-schema-list: CMD = schema list
-differ-schema-list: differ
-
-UBUNTU_BASE_IMAGE := focal-20220426
-SPARK_PSEUDO_BASE_IMAGE := 3.3.2-3.2.1
+DOCKER_BUILDX_BUILDER := multiarch
+SPARK_PSEUDO_BASE_IMAGE := 3.3.4-3.3.1
 MAKESTER__BUILD_COMMAND := --rm --no-cache\
+ --builder $(DOCKER_BUILDX_BUILDER)\
+ --platform $(DOCKER_PLATFORM)\
+ --build-arg RELEASE_VERSION=$(MAKESTER__RELEASE_VERSION)\
  --build-arg UBUNTU_BASE_IMAGE=$(UBUNTU_BASE_IMAGE)\
  --build-arg SPARK_PSEUDO_BASE_IMAGE=$(SPARK_PSEUDO_BASE_IMAGE)\
- -t $(MAKESTER__IMAGE_TAG_ALIAS) -f docker/Dockerfile .
-image-build: dep-package
+ --load\
+ --tag $(MAKESTER__IMAGE_TAG_ALIAS)\
+ -f docker/Dockerfile .
+
+diffit-image-build: gitversion-release py-distribution dep-package image-buildx
 
 ifndef SCHEMA
-SCHEMA := Dummy
-endif
-ifndef LEFT
-LEFT := file:///tmp/data/left
-endif
-ifndef RIGHT
-RIGHT := file:///tmp/data/right
+SCHEMA := $(PWD)/docker/files/schema
 endif
 ifndef DATA
 DATA := $(PWD)/docker/files/data
@@ -91,9 +87,9 @@ endif
 ifndef OUTPUT_PATH
 OUTPUT_PATH := file:///tmp/data/out
 endif
-MAKESTER__RUN_COMMAND = $(DOCKER) run\
- $(DOCKER_PLATFORM)\
+MAKESTER__RUN_COMMAND = $(MAKESTER__DOCKER) run\
  --rm -d\
+ --platform $(DOCKER_PLATFORM)\
  --publish 8032:8032\
  --publish 7077:7077\
  --publish 8080:8080\
@@ -103,63 +99,30 @@ MAKESTER__RUN_COMMAND = $(DOCKER) run\
  --env CORE_SITE__HADOOP_TMP_DIR=/tmp\
  --env HDFS_SITE__DFS_REPLICATION=1\
  --env YARN_SITE__YARN_NODEMANAGER_RESOURCE_DETECT_HARDWARE_CAPABILITIES=true\
- --env DIFFER_SCHEMA=$(SCHEMA)\
- --env LEFT_DATA_SOURCE=$(LEFT)\
- --env RIGHT_DATA_SOURCE=$(RIGHT)\
+ --env YARN_SITE__YARN_LOG_AGGREGATION_ENABLE=true\
  --env OUTPUT_PATH=$(OUTPUT_PATH)\
- --env AWS_ACCESS_KEY_ID=$(AWS_ACCESS_KEY_ID)\
- --env AWS_SECRET_ACCESS_KEY=$(AWS_SECRET_ACCESS_KEY)\
- --env AWS_SESSION_TOKEN=$(AWS_SESSION_TOKEN)\
  --env NUM_EXECUTORS=6\
  --env EXECUTOR_CORES=2\
  --env EXECUTOR_MEMORY=768m\
- --env COLS_TO_DROP="$(COLS_TO_DROP)"\
- --env RANGE_FILTER_COLUMN="$(RANGE_FILTER_COLUMN)"\
- --env RANGE_FILTER_LOWER="$(RANGE_FILTER_LOWER)"\
- --env RANGE_FILTER_UPPER="$(RANGE_FILTER_UPPER)"\
- --env RANGE_FILTER_FORCE="$(RANGE_FILTER_FORCE)"\
- --volume $(HOME)/.aws:/home/hdfs/.aws\
  --volume $(PWD)/docker/files/python:/data\
+ --volume $(SCHEMA):/tmp/schema\
  --volume $(DATA):/tmp/data\
  --hostname $(MAKESTER__CONTAINER_NAME)\
  --name $(MAKESTER__CONTAINER_NAME)\
- $(MAKESTER__IMAGE_TAG_ALIAS)
-
-ifndef JUPYTER_SERVER_PORT:
-JUPYTER_SERVER_PORT := 8889
-endif
-export JUPYTER_PORT = $(JUPYTER_SERVER_PORT)
-
-backoff:
-	@$(MAKESTER__PYTHON) makester/scripts/backoff -d "YARN ResourceManager" -p 8032 localhost
-	@$(MAKESTER__PYTHON) makester/scripts/backoff -d "YARN ResourceManager webapp UI" -p 8088 localhost
-	@$(MAKESTER__PYTHON) makester/scripts/backoff -d "YARN NodeManager webapp UI" -p 8042 localhost
-	@$(MAKESTER__PYTHON) makester/scripts/backoff -d "Spark HistoryServer web UI port" -p 18080 localhost
-	@$(MAKESTER__PYTHON) makester/scripts/backoff -d "Jupyter dashboard" -p $(JUPYTER_PORT) localhost
+ $(MAKESTER__IMAGE_TAG_ALIAS) $(CMD)
 
 MAKESTER__COMPOSE_FILES = -f docker/docker-compose.yml
-
-stack-up stack-down:
-stack-up: compose-up backoff stack-server
-
-stack-down: compose-down
-
-stack-server:
-	@$(DOCKER) exec -ti diffit-jupyter bash -c "jupyter notebook list"
 
 pyspark:
 	@PYSPARK_PYTHON=$(MAKESTER__PYTHON) pyspark --driver-memory=2G --conf spark.sql.session.timeZone=UTC
 
 help: makester-help docker-help py-help versioning-help
 	@echo "(Makefile)\n\
-  differ-schema-list   Show the differ tool schemas\n\
+  diffit-image-build   Diffit tooling container image builder\n\
   init                 Build the local Python-based virtual environment\n\
   init-dev             Build the local Python-based virtual environment (development)\n\
   lint                 Lint the code base\n\
   pyspark              Start the PyPI pyspark interpreter in virtual env context\n\
-  stack-down           Destroy local Jupyter Notebook server infrastructure\n\
-  stack-server         Get local Jupyter Notebook server URL\n\
-  stack-up             Create local Jupyter Notebook server infrastructure and intialisation\n\
   tests                Run code test suite\n"
 
 .PHONY: help tests
