+++ README.md
@@ -8,12 +8,6 @@
   - [Local Environment Maintenance](#Local-Environment-Maintenance)
 - [Help](#Help)
 - [Running the Test Harness](#Running-the-Test-Harness)
-- [Docker Image Management](#Docker-Image-Management)
-  - [Image Build](#Image-Build)
-  - [Validate the Image Build](#Validate-the-Image-Build)
-  - [Image Searches](#Image-Searches)
-  - [Docker Image Version Tag](#Docker-Image-Version-Tag)
-- [Useful `make` Commands](#Useful-make-Commands)
 - [FAQs](#FAQs)
 
 ## Overview
@@ -68,42 +62,6 @@ We use [pytest](https://docs.pytest.org/en/latest/). To run the tests:
 make tests
 ```
 
-## Docker Image Management
-### Image Build
-To build the Docker image against Apache Spark version `SPARK_VERSION` (defaults to `3.2.1`):
-```
-make build-image
-```
-You can target a specific Apache Spark release by setting `SPARK_VERSION`. For example:
-```
-SPARK_VERSION=3.0.3 make build image
-```
-The  tag naming policy follows `<spark-version>-<diffit-version>-<image-release-number>`.
-
-### Validate the Image Build
-A convenience target is provided for you to shake-out the new Docker image build against the `Dummy` schema:
-```
-make run
-```
-You can view the status of the job with:
-```
-make container-logs
-```
-On successful completion the `diffit` output will be written to `docker/files/data/out`
-
-### Image Searches
-Search for existing Docker image tags with command:
-```
-make image-search
-```
-### Docker Image Version Tag
-```
-make image-tag-version
-```
-Tagging convention used is:
-```
-<SBT_VERSION>-<RELEASE_VERSION>-<MAKESTER__RELEASE_NUMBER>
-```
 ## FAQs
 **_Q. Why do I get `WARNING: An illegal reflective access operation has occurred`?_**
 Seems to be related to the JVM version being used. Java 8 will suppress the warning. To check available Java versions on your Mac try `/usr/libexec/java_home -V`. Then:
