+++ setup.cfg
@@ -28,10 +28,10 @@ package_dir =
     =src
 install_requires =
     filester
-    pyspark==3.*
 
 [options.extras_require]
 dev =
+    pyspark==3.*
     pytest
     pytest-cov
     pytest-sugar
@@ -42,6 +42,7 @@ where = src
 [options.package_data]
 diffit =
     VERSION
+    schema/data/*.json
 
 [options.entry_points]
 console_scripts =
