+++ src/diffit/VERSION
@@ -1,1 +1,1 @@
-0.1.2b3
+0.1.2a5
